pyproject.toml
[tool.poetry]
name = "python-docuvortex-local"
version = "0.1.0"
description = "This project aims to implement a document-based question-answering system using a local LLM model, Python, and the Langchain Framework."
authors = ["Patrick Kalkman <patrick@simpletechture.nl>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.177"
python-dotenv = "^1.0.0"
tiktoken = "^0.4.0"
chromadb = "^0.3.25"
pdfplumber = "^0.9.0"
loguru = "^0.7.0"
pypdf = "^3.9.0"
streamlit = "^1.22.0"
streamlit-chat = "^0.0.2.2"
streamlit-extras = "^0.2.7"
pillow = "^9.5.0"
openai = "^0.27.7"
sentence-transformers = "^2.2.2"
instructorembedding = "^1.0.1"
accelerate = "^0.19.0"
bitsandbytes = "^0.39.0"
sentencepiece = "^0.1.99"
torch = "^2.0.1"
torchvision = "^0.15.2"
transformers = "^4.29.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
bandit = "^1.7.5"
pytest = "^7.3.1"

[tool.bandit]
exclude_dirs = ["tests",]

[tool.pytest.ini_options]
filterwarnings = ["ignore::Warning"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
